requirements.txt
aiofiles==23.2.1
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
colorzero==2.0
comm==0.2.0
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
coverage==7.3.2
cssselect2==0.7.0
cutlet==0.3.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.5
dataclass-factory==2.16
dateparser==1.1.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.7.0
encodec==0.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
fugashi==1.3.0
g2pkk==0.1.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
gpiozero==2.0.1
greenlet==3.0.3
grpcio==1.59.3
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.19.4
idna==3.5
inflect==7.0.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.18.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.22
jsonlines==1.2.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kaleido==0.2.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy-object-proxy==1.9.0
lazy_loader==0.3
librosa==0.10.0
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mecab-python3==1.0.6
mistune==3.0.2
mojimoji==0.0.12
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
neovim==0.3.1
nest-asyncio==1.5.8
networkx==2.8.8
nltk==3.8.1
nose2==0.14.0
notebook==7.1.1
notebook_shim==0.2.4
num2words==0.5.13
numba==0.58.1
numpy==1.26.4
oauthlib==3.2.2
overrides==7.7.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.4.0
pnglatex==1.1
pooch==1.8.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pydantic==1.10.15
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pylint==2.10.2
pynndescent==0.5.11
pynvim==0.5.0
pyparsing==3.1.1
pypinyin==0.49.0
pysbd==0.3.4
pytest==7.4.4
pytest-asyncio==0.20.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
shazamio==0.6.0
shazamio_core==1.0.7
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
SudachiDict-core==20230927
SudachiPy==0.6.7
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tkinter-async-execute==1.2.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
tornado==6.3.3
tqdm==4.66.1
trainer==0.0.32
traitlets==5.13.0
transformers==4.35.2
-e git+https://github.com/coqui-ai/TTS@11ec9f7471620ebaa57db7ff5705254829ffe516#egg=TTS
typer==0.9.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.7
unidic-lite==1.0.8
uri-template==1.3.0
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.12.1
yarl==1.9.3
